% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwayforpay-package.R
\docType{package}
\name{rwayforpay-package}
\alias{rwayforpay}
\alias{rwayforpay-package}
\title{rwayforpay: Get Transactions List From 'WayForPay'}
\description{
Provide function for work with 'WayForPay API transaction list', see \url{https://wiki.wayforpay.com/en/view/1736786}.
}
\author{
\strong{Maintainer}: Alexey Seleznev \email{selesnow@gmail.com} (\href{https://orcid.org/0000-0003-0410-7385}{ORCID})

}
\keyword{internal}
