% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfp_get_currency_rates.R
\name{wfp_get_currency_rates}
\alias{wfp_get_currency_rates}
\title{Get currency rates}
\usage{
wfp_get_currency_rates(
  date_begin = Sys.Date() - 10,
  date_end = Sys.Date() - 1,
  currency = NULL,
  merchant_login = wfp_get_merchant_login(),
  merchant_secret_key = wfp_get_secret_key(),
  api_version = 1
)
}
\arguments{
\item{date_begin}{start date}

\item{date_end}{end date}

\item{currency}{currency name}

\item{merchant_login}{Merchant's unique identifier.}

\item{merchant_secret_key}{Merchant's secret key}

\item{api_version}{protocol version. Default value: 1}
}
\value{
tibble with currency rates
}
\description{
Get currency rates
}
