% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfp_get_transactions_list.R
\name{wfp_get_transactions_list}
\alias{wfp_get_transactions_list}
\title{Get transactions list}
\usage{
wfp_get_transactions_list(
  date_begin = Sys.Date() - 30,
  date_end = Sys.Date() - 1,
  merchant_login = wfp_get_merchant_login(),
  merchant_secret_key = wfp_get_secret_key(),
  api_version = 2
)
}
\arguments{
\item{date_begin}{Period start date}

\item{date_end}{Period end date}

\item{merchant_login}{Merchant's unique identifier.}

\item{merchant_secret_key}{Merchant's secret key}

\item{api_version}{Protocol version. 2 - enable the transmission of extended data in the response - additional fields, delivery, comments.}
}
\value{
data.frame with transactions
}
\description{
Get transactions list
}
\seealso{
See \href{https://wiki.wayforpay.com/en/view/1736786}{documentation}
}
