% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfp_setters.R
\name{wfp_set_merchant_login}
\alias{wfp_set_merchant_login}
\title{Set Merchant Login}
\usage{
wfp_set_merchant_login(merchant_login)
}
\arguments{
\item{merchant_login}{merchant login}
}
\value{
not return value, use for set environment variable WFP_MERCHANT_LOGIN
}
\description{
Set Merchant Login
}
