% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfp_setters.R
\name{wfp_set_secret_key}
\alias{wfp_set_secret_key}
\title{Set Merchant Secret Key}
\usage{
wfp_set_secret_key(secret_key)
}
\arguments{
\item{secret_key}{merchant secret key}
}
\value{
not return value, use for set environment variable WFP_SECRET_KEY
}
\description{
Set Merchant Secret Key
}
